Ajustes

+++ database/migrations/2019_05_04_193920_create_detalle_ventas_table.php
@@ -20,7 +20,7 @@ class CreateDetalleVentasTable extends Migration
             $table->foreign('idventa')->references('id')->on('ventas')->onDelete('cascade');
             $table->integer('idproducto')->unsigned();
             $table->foreign('idproducto')->references('id')->on('productos');
-            $table->integer('cantidad');
+            $table->decimal('cantidad',11,2);
             $table->decimal('precio', 11, 2);
             $table->decimal('descuento',4, 2);
         });
